backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.0
pyagentspec==25.4.1
deepgram-sdk>=3.5.0
httpx>=0.25.2
pydantic>=2.10,<3
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
